custom-elements/custom-element-lifecycle/types-of-callbacks/entered-view-callback.html

<!DOCTYPE html>
<html>
<head>
<title>enteredView callback of a custom element</title>
<meta name="author" title="Sergey G. Grekhov" href="mailto:sgrekhov@unipro.ru">
<meta name="assert" content="enteredView callback must be enqueued whenever custom element is inserted into a document and this document has a browsing context">
<link rel="help" href="http://www.w3.org/TR/custom-elements/#types-of-callbacks">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src='../../testcommon.js'></script>
<link rel="stylesheet" href="/resources/testharness.css">
</head>
<body>
<div id="log"></div>
<script type="text/javascript">

test(function() {
    var doc = newHTMLDocument();
    var callbackCalledCounter = 0;

    var proto = Object.create(HTMLElement.prototype);
    proto.enteredViewCallback  = function () {
        callbackCalledCounter++;
    }
    var GeneratedConstructor = doc.register('x-a-1', {prototype: proto});
    var customElement = new GeneratedConstructor();
    assert_equals(callbackCalledCounter, 0, 'Callback enteredView was called for document ' +
            'that has no browsing content but should not');

}, 'Test enteredView callback when custom element is instantiating in JavaScript. ' +
    'Document has no browsing context');


test(function() {
    var doc = newHTMLDocument();
    var callbackCalledCounter = 0;

    var proto = Object.create(HTMLElement.prototype);
    proto.enteredViewCallback = function () {
        callbackCalledCounter++;
    }
    doc.register('x-a-2', {prototype: proto});
    doc.body.innerHTML = '<x-a-2></x-a-2>';
    assert_equals(callbackCalledCounter, 0, 'Callback enteredView was called for document ' +
            'that has no browsing content but should not');

}, 'Test enteredView callback when custom element is instantiating in HTML' +
    'Document has no browsing context');


test(function() {
    var doc = newHTMLDocument();
    var callbackCalledCounter = 0;

    var proto = Object.create(HTMLElement.prototype);
    proto.enteredViewCallback = function () {
        callbackCalledCounter++;
    }
    doc.body.innerHTML = '<x-a-3></x-a-3>';
    doc.register('x-a-3', {prototype: proto});
    assert_equals(callbackCalledCounter, 0, 'Callback enteredView was called for document ' +
            'that has no browsing content but should not');

}, 'Test enteredView callback when custom element is instantiating in HTML before ' +
    'registration of a custom element. Document has no browsing context');


test(function() {
    var doc = newHTMLDocument();
    var callbackCalledCounter = 0;

    doc.body.innerHTML = '<x-a-4 id="xa"></x-a-4>';
    var customElement = doc.querySelector('#xa');
    var proto = Object.create(HTMLElement.prototype);
    proto.enteredViewCallback = function () {
        callbackCalledCounter++;
    }
    customElement.constructor.prototype = proto;
    assert_equals(callbackCalledCounter, 0, 'Callback enteredView was called ' +
            'when custom element is uregistred and has no browsing context but should not');

}, 'Test enteredView callback when custom element is unregistered.');


testInIFrame('../../resources/x-element.html', function(context) {
    var doc = context.iframes[0].contentDocument;
    var callbackCalledCounter = 0;

    var proto = Object.create(HTMLElement.prototype);
    proto.enteredViewCallback = function () {
        callbackCalledCounter++;
    }
    doc.register('x-element', {prototype: proto});
    assert_equals(callbackCalledCounter, 1, 'Callback enteredView was not called for ' +
            'the document that has browsing content but should');

}, 'Test enteredView callback. Document has browsing context');


testInIFrame('../../resources/blank.html', function(context) {
    var doc = context.iframes[0].contentDocument;
    var callbackCalledCounter = 0;

    var proto = Object.create(HTMLElement.prototype);
    proto.enteredViewCallback = function () {
        callbackCalledCounter++;
    }
    doc.register('x-element', {prototype: proto});
    var x = doc.createElement('x-element');
    doc.body.appendChild(x);
    assert_equals(callbackCalledCounter, 1, 'Callback enteredView was not called for ' +
            'the document that has browsing content but should');

}, 'Test enteredView callback. Registred element is created via document.createElement(). ' +
    'Document has browsing context');


testInIFrame('../../resources/blank.html', function(context) {
    var doc = context.iframes[0].contentDocument;
    var callbackCalledCounter = 0;

    var x = doc.createElement('x-element');
    doc.body.appendChild(x);
    var proto = Object.create(HTMLElement.prototype);
    proto.enteredViewCallback = function () {
        callbackCalledCounter++;
    }
    doc.register('x-element', {prototype: proto});
    assert_equals(callbackCalledCounter, 1, 'Callback enteredView was not called for ' +
            'the document that has browsing content but should');

}, 'Test enteredView callback. Unregistred element is created via document.createElement(). ' +
    'Document has browsing context');


testInIFrame('../../resources/blank.html', function(context) {
    var doc = context.iframes[0].contentDocument;
    var callbackCalledCounter = 0;

    var proto = Object.create(HTMLElement.prototype);
    proto.enteredViewCallback = function () {
        callbackCalledCounter++;
    }
    doc.register('x-element', {prototype: proto});
    doc.body.innerHTML = '<x-element></x-element>';
    assert_equals(callbackCalledCounter, 1, 'Callback enteredView was not called for ' +
            'the document that has browsing content but should');

}, 'Test enteredView callback. Registered element is created via innerHTML property. ' +
    'Document has browsing context');


testInIFrame('../../resources/blank.html', function(context) {
    var doc = context.iframes[0].contentDocument;
    var callbackCalledCounter = 0;

    doc.body.innerHTML = '<x-element></x-element>';
    var proto = Object.create(HTMLElement.prototype);
    proto.enteredViewCallback = function () {
        callbackCalledCounter++;
    }
    doc.register('x-element', {prototype: proto});
    assert_equals(callbackCalledCounter, 1, 'Callback enteredView was not called for ' +
            'the document that has browsing content but should');

}, 'Test enteredView callback. Unresolved element is created via innerHTML property. ' +
    'Document has browsing context');

</script>
</body>
</html>
